build: add tests for building docker image (#3384)
* build: add tests for building docker image

Issue #2138 #2136 reference difficulties faced when our `docker build`
is broken. The commit attempts to mitigate the difficulties by adding a
sanity test in `make check` command. `make check` command is run in the
github workflow as well, so we would be able to locate any issues caused
due to a broken `docker build` quickly.

Signed-off-by: AnubhavUjjawal <anubhavujjawal@gmail.com>

* build: update docker-build-test

`docker-build-test` now exits with status code 0 if docker isn't installed
or there is an issue with the docker installation on the system.

Signed-off-by: AnubhavUjjawal <anubhavujjawal@gmail.com>

* build: changed exit code in docker-build-test to 1

Signed-off-by: AnubhavUjjawal <anubhavujjawal@gmail.com>

* build: change error msg and exit code of docker-build-test

Signed-off-by: AnubhavUjjawal <anubhavujjawal@gmail.com>

* build: change error msg of docker-build-test

Signed-off-by: AnubhavUjjawal <anubhavujjawal@gmail.com>

Co-authored-by: Ti Chi Robot <71242396+ti-chi-bot@users.noreply.github.com>