unsafe recovery: Add newly created empty regions output for unsafe recovery (#6869)
ref tikv/pd#6715

In DR auto sync scenario, it is possible that a empty range hole is left in the secondary data center 
when the primary data center is totally unavailable and the replication mode is "sync recover" (super 
rare). In this case, unsafe recovery + flashback is not able to recover the cluster. The fallback 
is BR + PiTR.  While, to help the people faster identify whether the fallback is needed, we need to add 
empty regions info in the output of unsafe recovery.

If there are empty regions created, use BR+PiTR. If there are no empty regions, use flashback.

Signed-off-by: Yang Zhang <yang.zhang@pingcap.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>