tests: make `TestBalanceRegion` more stable (#7036)
close tikv/pd#7035

seperate leader and peer scatter into two test. 
In the `TestBalanceRegion`, there are some case can't satisfy two constraint include leader and peer such as:
scatter 600 regions in 6 stores , and in the 600 selected, picked 1,3,4 as peer, but the onlt store 6 lead selected is fewest(99), so it don't select store 6 as leader.

Signed-off-by: bufferflies <1045931706@qq.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>