scheduler: refine the interval of scheduling tick in evict-slow-trend-scheduler. (#7326)
ref tikv/pd#7156

Implement the `GetNextInterval` for `evict-slow-trend-scheduler`, to refine the ticking interval.

Default `GetNextInterval` is not appropriate for `evict-slow-trend-scheduler`, as it might delay
the checking of other nodes' slowness status.

This pr adjusts the ticking interval of the evict-slow-trend-scheduler to optimize
its behavior. If a slow node is already identified as a candidate, the next
interval is now set to be shorter, ensuring quicker subsequent scheduling. 
This refinement aims to decrease response time.

Signed-off-by: lucasliang <nkcs_lykx@hotmail.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>